Re-organised directory structure. Cleanup.
