--- config/clickhouse/carbon-clickhouse.conf
@@ -1,82 +1,0 @@
-[common]
-# Prefix for store all internal carbon-clickhouse graphs. Supported macroses: {host}
-metric-prefix = "carbon.agents.{host}"
-# Endpoint for store internal carbon metrics. Valid values: "" or "local", "tcp://host:port", "udp://host:port"
-metric-endpoint = "local"
-# Interval of storing internal metrics. Like CARBON_METRIC_INTERVAL
-metric-interval = "1m0s"
-# GOMAXPROCS
-max-cpu = 1
-
-[logging]
-# "stderr", "stdout" can be used as file name
-file = "stdout"
-# Logging error level. Valid values: "debug", "info", "warn" "error"
-level = "warn"
-
-[data]
-# Folder for buffering received data
-path = "/data/carbon-clickhouse/"
-# Rotate (and upload) file interval.
-# Minimize chunk-interval for minimize lag between point receive and store
-chunk-interval = "1s"
-# Auto-increase chunk interval if the number of unprocessed files is grown
-# Sample, set chunk interval to 10 if unhandled files count >= 5 and set to 60s if unhandled files count >= 20:
-# chunk-auto-interval = "5:10s,20:60s"
-chunk-auto-interval = "5:5s,10:60s"
-
-[upload.graphite_reverse]
-type = "points-reverse"
-table = "graphite_reverse"
-threads = 1
-url = "http://clickhouse:8123/"
-timeout = "1m0s"
-
-[upload.graphite_tree]
-type = "tree"
-table = "graphite_tree"
-date = "2016-11-01"
-threads = 1
-url = "http://clickhouse:8123/"
-timeout = "1m0s"
-cache-ttl = "12h0m0s"
-
-[upload.graphite_series_reverse]
-type = "series-reverse"
-table = "graphite_series_reverse"
-threads = 1
-url = "http://clickhouse:8123/"
-timeout = "1m0s"
-
-[upload.graphite_tagged]
-type = "tagged"
-table = "graphite_tagged"
-threads = 1
-url = "http://clickhouse:8123/"
-timeout = "1m0s"
-cache-ttl = "12h0m0s"
-
-[udp]
-listen = ":2003"
-enabled = true
-
-[tcp]
-listen = ":2003"
-enabled = true
-
-[pickle]
-listen = ":2004"
-enabled = true
-
-[prometheus]
-listen = ":2006"
-enabled = true
-
-# https://github.com/lomik/carbon-clickhouse/blob/master/grpc/carbon.proto
-[grpc]
-listen = ":2005"
-enabled = false
-
-[pprof]
-listen = "localhost:7007"
-enabled = false
