--- config/clickhouse/graphite-clickhouse.conf
@@ -1,23 +1,0 @@
-[common]
-listen = ":9090"
-max-cpu = 8
-
-[clickhouse]
-url = "http://clickhouse:8123/?max_query_size=2097152&readonly=2"
-data-table = ""
-tree-table = "graphite_tree"
-rollup-conf = "/etc/graphite-clickhouse/rollup.xml"
-data-timeout = "1m0s"
-tree-timeout = "1m0s"
-date-tree-table = "graphite_series_reverse"
-date-tree-table-version = 3
-tagged-table = "graphite_tagged"
-
-[[data-table]]
-table = "graphite_reverse"
-reverse = true
-rollup-conf = "/etc/graphite-clickhouse/rollup.xml"
-
-[logging]
-file = "stdout"
-level = "warn"
