--- docker-compose-clickhouse.yaml
@@ -1,62 +1,0 @@
-version: "3"
-
-services:
-  clickhouse:
-    image: yandex/clickhouse-server:1.1.54380
-    volumes:
-      - "./config/clickhouse/rollup.xml:/etc/clickhouse-server/config.d/rollup.xml"
-      - "./data/clickhouse/data:/var/lib/clickhouse/data"
-      - "./data/clickhouse/metadata:/var/lib/clickhouse/metadata"
-
-  clickhouse-init:
-    image: yandex/clickhouse-server:1.1.54380
-    volumes:
-      - "./config/clickhouse/init.sql:/init.sql"
-    entrypoint: |
-      bash -c '
-        sleep 5
-        export HOME=/var/lib/clickhouse/
-        cat /init.sql | clickhouse client --host clickhouse --multiquery'
-    links:
-      - clickhouse
-
-  carbon-clickhouse:
-    image: lomik/carbon-clickhouse:v0.9.1
-    volumes:
-      - "./data/carbon-clickhouse:/data/carbon-clickhouse"
-      - "./config/clickhouse/carbon-clickhouse.conf:/etc/carbon-clickhouse/carbon-clickhouse.conf"
-    ports:
-      - "2003:2003" # plain tcp
-      - "2003:2003/udp" # plain udp
-      - "2004:2004" # pickle
-      - "2006:2006" # prometheus remote write
-    links:
-      - clickhouse
-
-  graphite-clickhouse:
-    image: lomik/graphite-clickhouse:v0.8.1
-    volumes:
-      - "./config/clickhouse/rollup.xml:/etc/graphite-clickhouse/rollup.xml"
-      - "./config/clickhouse/graphite-clickhouse.conf:/etc/graphite-clickhouse/graphite-clickhouse.conf"
-    links:
-      - clickhouse
-    ports:
-      - "9090:9090"
-
-  zipper:
-    build:
-      dockerfile: docker/zipper/Dockerfile
-      context: .
-    volumes:
-      - "./config/clickhouse/carbonzipper.yaml:/etc/carbonzipper.yaml"
-    links:
-      - graphite-clickhouse
-
-  carbonapi:
-    build:
-      dockerfile: docker/carbonapi/Dockerfile
-      context: .
-    ports:
-      - "8081:8081"
-    links:
-      - zipper
