+++ pkg/app/carbonapi/app.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"expvar"
 	"fmt"
-	"go.uber.org/zap/zapcore"
 	"net"
 	"net/http"
 	"os"
@@ -15,19 +14,21 @@ import (
 	"time"
 	"unicode"
 
-	"github.com/bookingcom/carbonapi/blocker"
-	"github.com/bookingcom/carbonapi/cache"
-	"github.com/bookingcom/carbonapi/carbonapipb"
-	"github.com/bookingcom/carbonapi/cfg"
+	"go.uber.org/zap/zapcore"
+
 	"github.com/bookingcom/carbonapi/expr/functions"
 	"github.com/bookingcom/carbonapi/expr/functions/cairo/png"
-	"github.com/bookingcom/carbonapi/mstats"
-	"github.com/bookingcom/carbonapi/pathcache"
 	"github.com/bookingcom/carbonapi/pkg/backend"
 	bnet "github.com/bookingcom/carbonapi/pkg/backend/net"
+	"github.com/bookingcom/carbonapi/pkg/blocker"
+	"github.com/bookingcom/carbonapi/pkg/cache"
+	"github.com/bookingcom/carbonapi/pkg/carbonapipb"
+	"github.com/bookingcom/carbonapi/pkg/cfg"
+	"github.com/bookingcom/carbonapi/pkg/mstats"
 	"github.com/bookingcom/carbonapi/pkg/parser"
+	"github.com/bookingcom/carbonapi/pkg/pathcache"
 	"github.com/bookingcom/carbonapi/pkg/trace"
-	"github.com/bookingcom/carbonapi/util"
+	"github.com/bookingcom/carbonapi/pkg/util"
 
 	"github.com/facebookgo/grace/gracehttp"
 	"github.com/facebookgo/pidfile"
