+++ pkg/pathcache/pathcache.go
